requirements.txt
Django==1.9
Fabric==1.10.2
MySQL-python==1.2.5
PyYAML==3.11
SQLAlchemy==1.0.11
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
boto==2.38.0
csvkit==0.9.1
dbf==0.94.003
django-admin-bootstrapped==2.5.7
django-bakery==0.7.6
django-debug-toolbar==1.4
django-mass-edit==2.6
django-tastypie==0.13.0
docutils==0.12
ecdsa==0.13
haversine==0.4.5
jdcal==1.2
latimes-calculate==0.3.1
lockfile==0.12.2
openpyxl==2.2.0-b1
paramiko==1.16.0
pycrypto==2.6.1
python-dateutil==2.4.2
python-googlegeocoder==0.3.0
python-mimeparse==0.1.4
pytz==2015.7
redis==2.10.5
requests==2.9.1
six==1.10.0
slackbot==0.3.0
slacker==0.9.0
sqlparse==0.1.18
twitter==1.15.0
websocket-client==0.35.0
wsgiref==0.1.2
xlrd==0.9.4
